Separate telemetry state lifecycle
Keep the serialized private-state model in the module root and move
identifier reset, window, and cohort transitions into lifecycle.rs.

Leave each responsibility's tests beside its implementation and record
the boundary in the module structure guide.
